Fix active item on focus
and refactor tests for listbox
