+++ packages/aria-voyager/tests/listbox/selection/multi/keyboard/home.test.ts
@@ -1,0 +1,30 @@
+import { expect, test } from 'vitest';
+
+import { Listbox } from '../../../../../src';
+import { createMultiSelectListWithFruits } from '../../../-shared';
+
+test('select from third to first item with `Home` and `Shift` key', () => {
+  const list = createMultiSelectListWithFruits();
+
+  new Listbox(list);
+
+  const firstItem = list.children[0];
+  const secondItem = list.children[1];
+  const thirdItem = list.children[2];
+
+  expect(firstItem.getAttribute('aria-selected')).toBeNull();
+  expect(secondItem.getAttribute('aria-selected')).toBeNull();
+  expect(thirdItem.getAttribute('aria-selected')).toBeNull();
+
+  thirdItem.dispatchEvent(new PointerEvent('pointerup', { bubbles: true }));
+
+  expect(firstItem.getAttribute('aria-selected')).toBeNull();
+  expect(secondItem.getAttribute('aria-selected')).toBeNull();
+  expect(thirdItem.getAttribute('aria-selected')).toBe('true');
+
+  list.dispatchEvent(new KeyboardEvent('keydown', { key: 'Home', shiftKey: true }));
+
+  expect(firstItem.getAttribute('aria-selected')).toBe('true');
+  expect(secondItem.getAttribute('aria-selected')).toBe('true');
+  expect(thirdItem.getAttribute('aria-selected')).toBe('true');
+});
