+++ packages/aria-voyager/tests/listbox/selection/single/keyboard/end.test.ts
@@ -1,0 +1,24 @@
+import { expect, test } from 'vitest';
+
+import { Listbox } from '../../../../../src';
+import { createListWithFruits } from '../../../-shared';
+
+test('select last item with `End` key', () => {
+  const list = createListWithFruits();
+
+  new Listbox(list);
+
+  const firstItem = list.children[0];
+  const secondItem = list.children[1];
+  const thirdItem = list.children[2];
+
+  expect(firstItem.getAttribute('aria-selected')).toBeNull();
+  expect(secondItem.getAttribute('aria-selected')).toBeNull();
+  expect(thirdItem.getAttribute('aria-selected')).toBeNull();
+
+  list.dispatchEvent(new KeyboardEvent('keydown', { key: 'End' }));
+
+  expect(firstItem.getAttribute('aria-selected')).toBeNull();
+  expect(secondItem.getAttribute('aria-selected')).toBeNull();
+  expect(thirdItem.getAttribute('aria-selected')).toBe('true');
+});
