+++ packages/aria-voyager/tests/listbox/selection/single/keyboard/home.test.ts
@@ -1,0 +1,24 @@
+import { expect, test } from 'vitest';
+
+import { Listbox } from '../../../../../src';
+import { createListWithFruits } from '../../../-shared';
+
+test('select first item with `Home` key', () => {
+  const list = createListWithFruits();
+
+  new Listbox(list);
+
+  const firstItem = list.children[0];
+  const secondItem = list.children[1];
+  const thirdItem = list.children[2];
+
+  expect(firstItem.getAttribute('aria-selected')).toBeNull();
+  expect(secondItem.getAttribute('aria-selected')).toBeNull();
+  expect(thirdItem.getAttribute('aria-selected')).toBeNull();
+
+  list.dispatchEvent(new KeyboardEvent('keydown', { key: 'Home' }));
+
+  expect(firstItem.getAttribute('aria-selected')).toBe('true');
+  expect(secondItem.getAttribute('aria-selected')).toBeNull();
+  expect(thirdItem.getAttribute('aria-selected')).toBeNull();
+});
